+++ backend/src/main/java/com/stemlink/skillmentor/services/StudentService.java
@@ -8,8 +8,12 @@ import java.util.List;
 public interface StudentService {
 
     Student createNewStudent(Student student);
+
     List<Student> getAllStudents();
-    Student getStudentById(Integer id);
-    Student updateStudentById(Integer id, Student updatedStudent);
-    void deleteStudent(Integer id);
+
+    Student getStudentById(Long id);
+
+    Student updateStudentById(Long id, Student updatedStudent);
+
+    void deleteStudent(Long id);
 }
